requirements.txt
apache-airflow==2.7.0
psycopg2-binary==2.9.6
pandas==1.5.3
numpy==1.24.3
python-dotenv==1.0.0
pendulum==2.1.2


# Dependencias para DGBienes
oracledb==1.4.2
openpyxl==3.1.2
pyodbc==5.0.1

# Dependencias adicionales si las usas
sqlalchemy